+++ examples/src/02-pixi-intro/primitives.ts
@@ -1,22 +1,13 @@
 import * as PIXI from 'pixi.js';
+import { PIXIExample } from '../utils/APHExample';
+export class Primitives extends PIXIExample {
 
-export class PrimitivesExample extends PIXI.Application {
-
-	constructor(view: HTMLCanvasElement) {
-		super({
-            view,
-			backgroundColor: 0x000000,
-			width: view.clientWidth,
-			height: view.clientHeight,
-		});
-
-		this.init();
-	}
-
-	init() {
+	load() {
         // TODO put your code here - create PIXI.Graphics()
         // don't forget to add the root object to this.stage 
 	}
-}
 
-new PrimitivesExample(<HTMLCanvasElement>document.getElementById('gameCanvas'));
+	update() {
+		// no-op
+	}
+}